config for datasource and migration
