+++ backend/.gitignore
@@ -2,3 +2,5 @@ node_modules/
 /package-lock.json
 /.env
 /ormconfig.json
+/dist
+note.md
