+++ backend/linkapi/favorite.http
@@ -1,6 +1,7 @@
 GET http://localhost:3000/api/favorite
 
-
+###
+GET http://localhost:3000/
 ###
 GET http://localhost:3000/api/favorite/1
