+++ backend/package.json
@@ -5,16 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon --exec ts-node src/index.ts",
-    "typeormold": "node --require ts-node/register ./node_modules/typeorm/cli.js",
-    "typeorm": "ts-node -P ./tsconfig.typeorm.json -d ./src/db/CliDataSource.ts",
-
-    "migration:run": "ts-node ./node_modules/typeorm/cli.js migration:run -d ./src/datasource.ts",
-    "migration:revert": "ts-node ./node_modules/typeorm/cli.js migration:revert -d ./src/datasource.ts",
-    "schema:sync": "",
-    "migration:show": "npm run typeorm migration:show -- -d ./src/datasource.ts",
-    "migration:generate": "npm run typeorm migration:generate -- -d ./src/datasource.ts",
-    
-    "migration:create": "npm run typeorm migration:create"
+    "typeorm": "typeorm-ts-node-commonjs"
   },
   "keywords": [],
   "author": "",
