--- backend/src/entities/User.entities.ts
@@ -1,22 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, ManyToOne, ManyToMany, JoinTable } from 'typeorm';
-import { Role } from './Role.entities';
-import { Favorite } from './Favorite.entities';
-
-@Entity()
-export class User {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  username: string;
-
-  @Column()
-  password: string;
-
-  @ManyToOne(() => Role, role => role.users)
-  role: Role;
-
-  @ManyToMany(() => Favorite)
-  @JoinTable()
-  favorites: Favorite[];
-}
