+++ backend/src/index.ts
@@ -1,13 +1,12 @@
+import 'reflect-metadata'
 import express from 'express';
 import router from './routes';
-import { connectToDatabase } from './connection';
 import { dataSource } from './datasource';
 const app = express();
 const port = process.env.PORT || 3000;
 
 app.use(express.json());
 
-dataSource
 app.use('/api', router)
 
 app.get('/', (req, res) => {
